Improved User logic

+++ Auktionssajt.Core/Services/UserService.cs
@@ -1,58 +1,66 @@
 using Auktionssajt.Core.Interfaces;
-using Auktionssajt.Data.Repository;
 using Auktionssajt.Domain.DTOs;
-using Auktionssajt.Domain.Entities;
 using Auktionssajt.Domain.Models;
-using Auktionssajt.Core.Services; 
-using System.Collections.Generic;
-using Auktionssajt.Data.Interfaces;
+using Auktionssajt.Data.Repository;
 
 namespace Auktionssajt.Core.Services
 {
     public class UserService : IUserService
     {
-        private readonly IUserRepo _userRepo;
-        private readonly MappingService _mappingService;
+        private readonly UserRepo _userRepo = new();
+        private readonly MappingService _mappingService = new();
+        private readonly ValidationService _validationService = new();
 
-        public UserService(IUserRepo userRepo, MappingService mappingService)
+        public Status NewUser(NewUserModel user)
         {
-            _userRepo = userRepo;
-            _mappingService = mappingService;
+            var status = _validationService.ValidateUser(user);
+            if (status != Status.Ok)
+                return status;
+
+            var entity = _mappingService.ToUserEntity(user);
+            _userRepo.NewUser(entity);
+
+            return Status.Ok;
         }
 
-        public void DeleteUser(int id)
+        public Status DeleteUser(int id, int userId)
         {
+            var user = _userRepo.GetUser(id);
+            if (user == null)
+                return Status.NotFound;
+            
+            if (user.UserID != userId)
+                return Status.Unauthorized;
+            
             _userRepo.DeleteUser(id);
+            return Status.Ok;
         }
 
         public List<UserDTO> GetAllUsers()
         {
             var entities = _userRepo.GetAllUsers();
-            return _mappingService.UserEntityToUserDTO(entities); 
+            if (entities.Count < 1)
+                return null!;
+
+            return _mappingService.ToUserDTO(entities);
         }
 
         public UserDTO GetUser(int id)
         {
             var entity = _userRepo.GetUser(id);
-            return _mappingService.UserEntityToUserDTO(entity);
+            if (entity == null)
+                return null!;
+            
+            return _mappingService.ToUserDTO(entity);
         }
 
         public UserDTO GetUser(string userName)
         {
             var entity = _userRepo.GetUser(userName);
-            return _mappingService.UserEntityToUserDTO(entity);
-        }
+            if (entity == null)
+                return null!;
 
-        public void NewUser(NewUserModel user)
-        {
-            var entity = _mappingService.NewUserModelToUserEntity(user);
-            _userRepo.NewUser(entity);
-        }
-
-        public void UpdateUser(UpdateUserModel user)
-        {
-            var entity = _mappingService.UpdateUserModelToUserEntity(user);
-            _userRepo.UpdateUser(entity);
+            return _mappingService.ToUserDTO(entity);
         }
     }
 }
